finished API ref description. fixed bugs, deleted error as a return var in CreatePool. Changed Job.state.UpdateData and Job.state.GetStateData. Now it's working only with state's Data

+++ internal/domain/retry.go
@@ -7,7 +7,7 @@ type Retry struct {
 
 	// Count is the number of allowed retry attempts before marking the job as failed.
 	// Each time a job fails, this counter decreases. When it reaches zero, the job stops retrying.
-	Count int64
+	Count int
 
 	// ResetOnSuccess determines whether the retry counter should be reset after a successful execution.
 	// If true, a successful execution resets the retry count to its initial value.
